refactor useWindowSize as hook with JSDoc
