+++ src/lib/hooks/useWindowSize.ts
@@ -1,0 +1,30 @@
+import { useEffect, useState } from 'react';
+
+// Define general type for useWindowSize hook, which includes width and height
+export interface Size {
+  width: number | undefined;
+  height: number | undefined;
+}
+/**
+ * A custom React hook that returns the current window size.
+ *
+ * @returns An object with `width` and `height` properties representing the current window size.
+ */
+export function useWindowSize(): Size {
+  const [windowSize, setWindowSize] = useState<Size>({
+    width: undefined,
+    height: undefined,
+  });
+  useEffect(() => {
+    function handleResize() {
+      setWindowSize({
+        width: window.innerWidth,
+        height: window.innerHeight,
+      });
+    }
+    window.addEventListener('resize', handleResize);
+    handleResize();
+    return () => window.removeEventListener('resize', handleResize);
+  }, []);
+  return windowSize;
+}
