+++ src/lib/utils.ts
@@ -1,5 +1,4 @@
 import type { CustomErrorBody, ValidatorError } from '@/lib/types/apiResponses';
-import { useEffect, useState } from 'react';
 
 /**
  * Get next `num` years from today in a number array to generate dropdown options for future selections
@@ -33,23 +32,4 @@ export const getMessagesFromError = (errBody: CustomErrorBody): string[] => {
 export interface Size {
   width: number | undefined;
   height: number | undefined;
-}
-
-export function useWindowSize(): Size {
-  const [windowSize, setWindowSize] = useState<Size>({
-    width: undefined,
-    height: undefined,
-  });
-  useEffect(() => {
-    function handleResize() {
-      setWindowSize({
-        width: window.innerWidth,
-        height: window.innerHeight,
-      });
-    }
-    window.addEventListener('resize', handleResize);
-    handleResize();
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-  return windowSize;
 }
